Rename tls_handshake; update docs
